refactor: Move back to single file
